+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+---
+
+## [Unreleased]
+
+### Fixed
+- When first starting, the memory buckets are not all created. The means that for the first few games the bucket_counts may be unexpectly short, this causes issues on the client side in snapshot report, which expects the data to be a specific size. The simple fix is to have the `ATHDataMgr` check that the bucket count is correct before sending the metrics data.
+
 ---
 
 ## [0.23.1] - 2026-03-28 @ 11:38 - HOTFIX